requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
yfinance
google-generativeai
python-dotenv